requirements.txt
appdirs==1.4.3
distlib==0.3.1
filelock==3.0.12
identify==1.4.14
pipenv==2020.11.15
pygerduty==0.38.3
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.14.0
virtualenv==20.2.1
virtualenv-clone==0.5.4
